refactor: separate info pane components
