+++ components/InfoPane/FileUploadControl.scss
