+++ components/InfoPane/FileUploadList.scss
