+++ components/QueryLink.tsx
@@ -1,9 +1,15 @@
 import filterAlteredClicks from 'filter-altered-clicks';
-import React from 'react';
+import React, { HTMLProps } from 'react';
 import useLocation from '../lib/useLocation.js';
 import { useQuery } from './App.js';
 
-export function QueryLink({ query }: { query: string | string[] }) {
+export function QueryLink({
+  query,
+  children,
+  ...props
+}: HTMLProps<HTMLAnchorElement> & {
+  query: string | string[];
+}) {
   const [, setQuery] = useQuery();
   const [location] = useLocation();
 
@@ -17,9 +23,17 @@ export function QueryLink({ query }: { query: string | string[] }) {
   const url = new URL(location);
   url.search = query.length ? `q=${queries.join(',')}` : '';
 
+  if (!children) {
+    return (
+      <a href={url.href} onClick={filterAlteredClicks(onClick)} {...props}>
+        {queries.join(',')}
+      </a>
+    );
+  }
+
   return (
-    <a href={url.href} onClick={filterAlteredClicks(onClick)}>
-      {queries.join(',')}
+    <a href={url.href} onClick={filterAlteredClicks(onClick)} {...props}>
+      {children}
     </a>
   );
 }
