+++ index.html
@@ -12,6 +12,7 @@
       ws://localhost:1234
       *.doubleclick.net
       api.github.com
+      raw.githubusercontent.com
       api.npmjs.org
       api.npms.io
       bundlephobia.com
